host_app/requirements.txt
google-adk==1.2.1
a2a-sdk
nest_asyncio
python-dotenv
asyncclick
PyJWT
jwcrypto
uvicorn
